remove cn.ts and add zh.ts
per request, change the name and locale setup
